// src/declarations/file_manager/file_manager.did.d.ts
export type Asset = { File: FileResponse } | { Directory: DirectoryResponse };
// ...
export interface NestedAssets {
  asset: PostAsset;
  children: Vec_1;
}
export type Permission = { Private: null } | { Origin: Array<string> } | { Public: null };
export type PostAsset = { File: PostFile } | { None: null } | { Directory: PostDirectory };
export interface PostDirectory {
  permission: Permission;
  name: string;
  children: Array<Asset>;
  parent_id: [] | [bigint];
}
export interface PostFile {
  permission: Permission;
  origin_path: string;
  metadata: [] | [string];
  name: string;
  size: bigint;
  mime_type: string;
  parent_id: [] | [bigint];
  chunk_count: bigint;
  extension: string;
}
// ...
export type Vec = Array<
  | {
      File: {
        id: bigint;
        permission: { Private: null } | { Origin: Array<string> } | { Public: null };
        updated_at: bigint;
        sha256: Uint8Array | number[];
        owner: [] | [Principal];
        metadata: [] | [string];
        name: string;
        path: string;
        size: bigint;
        mime_type: string;
        created_at: bigint;
        parent_id: [] | [bigint];
        chunks: BigUint64Array | bigint[];
        is_protected: boolean;
        extension: string;
      };
    }
  | {
      Directory: {
        id: bigint;
        permission: { Private: null } | { Origin: Array<string> } | { Public: null };
        updated_at: bigint;
        owner: [] | [Principal];
        name: string;
        created_at: bigint;
        children: Vec;
        parent_id: [] | [bigint];
        is_protected: boolean;
      };
    }
>;
export type Vec_1 = Array<{
  asset:
    | {
        File: {
          permission: { Private: null } | { Origin: Array<string> } | { Public: null };
          origin_path: string;
          metadata: [] | [string];
          name: string;
          size: bigint;
          mime_type: string;
          parent_id: [] | [bigint];
          chunk_count: bigint;
          extension: string;
        };
      }
    | { None: null }
    | {
        Directory: {
          permission: { Private: null } | { Origin: Array<string> } | { Public: null };
          name: string;
          children: Array<
            | {
                File: {
                  id: bigint;
                  permission: { Private: null } | { Origin: Array<string> } | { Public: null };
                  updated_at: bigint;
                  sha256: Uint8Array | number[];
                  owner: [] | [Principal];
                  metadata: [] | [string];
                  name: string;
                  path: string;
                  size: bigint;
                  mime_type: string;
                  created_at: bigint;
                  parent_id: [] | [bigint];
                  chunks: BigUint64Array | bigint[];
                  is_protected: boolean;
                  extension: string;
                };
              }
            | {
                Directory: {
                  id: bigint;
                  permission: { Private: null } | { Origin: Array<string> } | { Public: null };
                  updated_at: bigint;
                  owner: [] | [Principal];
                  name: string;
                  created_at: bigint;
                  children: Vec;
                  parent_id: [] | [bigint];
                  is_protected: boolean;
                };
              }
          >;
          parent_id: [] | [bigint];
        };
      };
  children: Vec_1;
}>;